requirements/base.txt
fastapi
asyncpg
uvicorn
sqlalchemy
alembic
python-multipart
python-dotenv
psycopg2-binary
langchain
langchain-openai
langchain_community
langgraph
tavily-python
langgraph-checkpoint-sqlite


